⏲🍵 Checkpoint ./public/client.js:1020861/87

+++ public/client.js
@@ -7,11 +7,14 @@ var init = function() {
   //get data from server
   axios.get("/data").then((response) => {
     let data = response.data
-    alert(data)
-    for (let i in data) {
-      let x = data[i]
-      let y = dateFormat(x.date, "dS mmm. yyyy")
-      html += `<tr><td>${y}</td><td>${x.country}</td><td>${x.place}</td><td>${x.notes}</td><td><a href="/delete?id=${x.id}">X</a></td></tr>`
+    if (!data) {
+      
+    } else {
+      for (let i in data) {
+        let x = data[i]
+        let y = dateFormat(x.date, "dS mmm. yyyy")
+        html += `<tr><td>${y}</td><td>${x.country}</td><td>${x.place}</td><td>${x.notes}</td><td><a href="/delete?id=${x.id}">X</a></td></tr>`
+      }
     }
     
     //update page
